bot-notifier/pyproject.toml
[tool.poetry]
name = "notifier"
version = "1.0"
description = ""
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
flask-cors = "^4.0.1"
waitress = "^3.0.0"
requests = "^2.32.3"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.6"
black = "^25.1.0"
mypy = "^1.15.0"
ruff = "^0.9.7"
types-pyyaml = "^6.0.12.20241230"
types-requests = "^2.32.0.20241016"
types-flask-cors = "^5.0.0.20240902"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
